Exposed protein allele model to public surface; moved shared allele variant modeling to the shared module.
